+++ src/alfred3/element/core.py
@@ -799,7 +799,7 @@ class Element:
         :class:`.JavaScript` element is a better choice.
 
         Args:
-            code: Css code
+            code: JavaScript code
             priority: Can be used to influence the order in which code
                 is added to the page. Sorting is ascending, i.e. the
                 lowest numbers appear closest to the top of the page.
@@ -1328,13 +1328,6 @@ class Row(Element):
 
             if self.elements_full_width:
                 element.width = "full"
-
-    # def _prepare_web_widget(self):
-
-    #     for element in self.elements:
-    #         element.prepare_web_widget()
-
-    #     super()._prepare_web_widget()
 
     @property
     def _cols(self) -> Iterator: